requirements.txt
numpy==1.21.4
numpy-stl==2.16.3
PyOpenGL==3.1.5
PySide2==5.15.2
pyqtgraph==0.12.3
